CD Planner Data Rework. More Bosses/Abilities Added
Moved Bosslist and Bossabilities to seperate files.
Added More Bosses
