+++ src/General/Modules/CooldownPlanner/Data/CooldownPlannerBossList.js
@@ -1,0 +1,142 @@
+export const bossList = [
+  /* ---------------------------------------------------------------------------------------------- */
+  /*                                            Nyalotha                                            */
+  /* ---------------------------------------------------------------------------------------------- */
+
+  {
+    DungeonEncounterID: 2344,
+    ID: 2375,
+    name: "N'zoth",
+    zoneID: 2217,
+  },
+
+  /* ---------------------------------------------------------------------------------------------- */
+  /*                                         Castle Nathria                                         */
+  /* ---------------------------------------------------------------------------------------------- */
+
+  {
+    DungeonEncounterID: 2398,
+    ID: 2393,
+    name: "Shriekwing",
+    zoneID: 2296,
+  },
+  {
+    DungeonEncounterID: 2418,
+    ID: 2429,
+    name: "Huntsman Altimor",
+    zoneID: 2296,
+  },
+  {
+    DungeonEncounterID: 2402,
+    ID: 2422,
+    name: "Sun King's Salvation",
+    zoneID: 2296,
+  },
+  {
+    DungeonEncounterID: 2405,
+    ID: 2418,
+    name: "Artificer Xy'mox",
+    zoneID: 2296,
+  },
+  {
+    DungeonEncounterID: 2383,
+    ID: 2428,
+    name: "Hungering Destroyer",
+    zoneID: 2296,
+  },
+  {
+    DungeonEncounterID: 2406,
+    ID: 2420,
+    name: "Lady Inerva Darkvein",
+    zoneID: 2296,
+  },
+  {
+    DungeonEncounterID: 2412,
+    ID: 2426,
+    name: "The Council of Blood",
+    zoneID: 2296,
+  },
+  {
+    DungeonEncounterID: 2399,
+    ID: 2394,
+    name: "Sludgefist",
+    zoneID: 2296,
+  },
+  {
+    DungeonEncounterID: 2417,
+    ID: 2425,
+    name: "Stone Legion Generals",
+    zoneID: 2296,
+  },
+  {
+    DungeonEncounterID: 2407,
+    ID: 2424,
+    name: "Sire Denathrius",
+    zoneID: 2296,
+  },
+
+  /* ---------------------------------------------------------------------------------------------- */
+  /*                                      Sanctum of Domination                                     */
+  /* ---------------------------------------------------------------------------------------------- */
+
+  {
+    DungeonEncounterID: 2423,
+    ID: 2435,
+    name: "The Tarragrue",
+    zoneID: 2450,
+  },
+  {
+    DungeonEncounterID: 2433,
+    ID: 2442,
+    name: "The Eye of the Jailer",
+    zoneID: 2450,
+  },
+  {
+    DungeonEncounterID: 2429,
+    ID: 2439,
+    name: "The Nine",
+    zoneID: 2450,
+  },
+  {
+    DungeonEncounterID: 2432,
+    ID: 2444,
+    name: "Remnant of Ner'zhul",
+    zoneID: 2450,
+  },
+  {
+    DungeonEncounterID: 2434,
+    ID: 2445,
+    name: "Soulrender Dormazain",
+    zoneID: 2450,
+  },
+  {
+    DungeonEncounterID: 2430,
+    ID: 2443,
+    name: "Painsmith Raznal",
+    zoneID: 2450,
+  },
+  {
+    DungeonEncounterID: 2436,
+    ID: 2446,
+    name: "Guardian of the First Ones",
+    zoneID: 2450,
+  },
+  {
+    DungeonEncounterID: 2431,
+    ID: 2447,
+    name: "Fatescribe Roh-Kalo",
+    zoneID: 2450,
+  },
+  {
+    DungeonEncounterID: 2422,
+    ID: 2440,
+    name: "Kel'Thuzad",
+    zoneID: 2450,
+  },
+  {
+    DungeonEncounterID: 2435,
+    ID: 2441,
+    name: "Sylvanas Windrunner",
+    zoneID: 2450,
+  },
+];
