+++ src/locale/en/translate.json
@@ -242,15 +242,15 @@
         "332626": "Shattering Pain",
         "326833": "Painful Memories",
         "338687": "Nathrian Hymn",
-        "SanctumOfDomination": "",
-        "TheTarragrue": "",
+        "Sanctum of Domination": "",
+        "The Tarragrue": "",
         "347668": "Grasp of Death",
         "352392": "Remnant: Soulforge Heat",
         "354173": "Chains of Eternity",
         "347737": "Hungering Mist",
         "346985": "Overpower",
         "347274": "Annihilating Smash",
-        "TheEyeOfTheJailer": "",
+        "The Eye of The Jailer": "",
         "355245": "Ire",
         "355240": "Scorn",
         "349028": "Titanic Death Gaze",
@@ -267,7 +267,7 @@
         "350809": "Jailer's Misery",
         "350847": "Desolation Beam",
         "350713": "Slothful Corruption",
-        "TheNine": "",
+        "The Nine": "",
         "351399": "Resentment",
         "350206": "Repeating Slash",
         "350287": "Song of Dissolution",
@@ -277,7 +277,47 @@
         "350486": "Link Essence",
         "350283": "Soulful Blast",
         "350475": "Pierce Soul",
-        "350098": "Aradne's Falling Strike"
+        "350098": "Aradne's Falling Strike",
+        "Remnant of Nerzhul": "",
+        "354632": "Aura of Spite",
+        "350073": "Torment",
+        "350427": "Malevolence",
+        "350489": "Lingering Malevolence",
+        "354534": "Spite",
+        "351073": "Shatter",
+        "349890": "Suffering",
+        "350076": "Grasp of Malice",
+        "Soulrender Dormazain": "",
+        "350801": "Unleashed Tyranny",
+        "354231": "Soul Manacles",
+        "353429": "Tormented",
+        "351787": "Agonizing Spike",
+        "350217": "Torment",
+        "351229": "Rendered Soul",
+        "350927": "Warmonger Shackle",
+        "350851": "Vessel of Torment",
+        "350422": "Ruinblade",
+        "Painsmith Raznal": "",
+        "355506": "Shadowsteel Chains",
+        "348463": "Flameclasp Detonation",
+        "355535": "Shadowsteel Embers",
+        "355786": "Blackened Armor",
+        "355788": "Dualblade Scythe",
+        "348255": "Spiked",
+        "348520": "Rippling Hammer",
+        "355573": "Cruciform Axe",
+        "348363": "Spiked Ball",
+        "Fatescribe Roh-Kalo":  "",
+        "351685": "Heroic Destiny",
+        "353931": "Twist Fate",
+        "353696": "Unstable Fate",
+        "350819": "Call of Eternity",
+        "353435": "Overwhelming Burden",
+        "353603": "Diviner's Probe",
+        "353433": "Burden of Destiny",
+        "353162": "Frayed Fate",
+        "350355": "Fated Conjunction",
+        "353398": "Anomalous Blast"
       }
     },
     "HallOfFame": {
